add cross message api
