+++ src/server/handlers/validator.rs
@@ -4,7 +4,7 @@
 
 use crate::config::ReloadableConfig;
 use crate::lotus::client::LotusJsonRPCClient;
-use crate::lotus::LotusClient;
+use crate::lotus::LotusIPCClient;
 use crate::server::JsonRPCRequestHandler;
 use anyhow::anyhow;
 use async_trait::async_trait;
@@ -58,7 +58,7 @@ impl JsonRPCRequestHandler for QueryValidatorSetHandler {
         };
 
         let lotus = LotusJsonRPCClient::from_subnet(subnet);
-        let response = lotus.ipc_read_subnet_actor_state(tip_set).await?;
+        let response = lotus.read_subnet_actor_state(tip_set).await?;
 
         Ok(QueryValidatorSetResponse {
             validator_set: response.validator_set,
